services/gmail_services.py
from googleapiclient.discovery import build
from google.oauth2.credentials import Credentials

def get_emails():
    creds = Credentials.from_authorized_user_file("token.json")
    service = build("gmail", "v1", credentials=creds)

    results = service.users().messages().list(userId='me', maxResults=10).execute()
    messages = results.get('messages', [])

    return messages
