requirements.txt
flask
transformers
torch
scikit-learn
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
requests
pandas
numpy
sentencepiece
accelerate
